updated documentation, focused on the synthetic data example
