+++ docs/examples/zhuang_lab_mouse_brain.md
@@ -13,7 +13,7 @@ We are only going to analyze one of their samples, specifically `mouse1_sample1_
 - [merfish3d-analysis](https://www.github.com/qi2lab/merfish3d-analysis)
 - [Baysor](https://github.com/kharchenkolab/Baysor)
 - [Raw 2D MERFISH data](https://download.brainimagelibrary.org/cf/1c/cf1c1a431ef8d021/)
-    - Download the `additional_files` folder, `mouse1_sample1_raw` folder, and .
+    - Download the `additional_files/` folder, `mouse1_sample1_raw/` folder, and `processed_data/spots_mouse1sample1.csv`.
 
 ## Downloading the data
 
@@ -32,7 +32,7 @@ All of the required code to process this data is in the [examples/zhuang_lab](ht
     ├── microscope.json
     ├── fov_positions/
       └── mouse1_sample1.txt
-  └── zhuang_decoded_codewords/
+  └── processed_data/
     └── spots_mouse1sample1.csv
 ```
 
@@ -40,7 +40,7 @@ All of the required code to process this data is in the [examples/zhuang_lab](ht
 
 Because this data is generated by a different custom microscope design with a unique microscope control package and is already pre-processed, we have to manually write most of the data conversion steps. Please see the [DataStore](../datastore.md) page for more information on the key parameters that are requred to create a `qi2labDataStore`.
 
-The data in this BIL dataset consists of six small section of mouse brain, all separated on the slide. Looking at the `fov_positions/mouse1_sample1.txt` file, we can find where there are discontinuous jumps in the stage positions. Those jumps indicate the stage moved to a new brain section. 
+The data in this BIL dataset consists of six small section of mouse brain, all separated on the slide. Looking at the `additional_files/fov_positions/mouse1_sample1.txt` file, we can find where there are discontinuous jumps in the stage positions. Those jumps indicate the stage moved to a new brain section. 
 
 A key issue with this data is the stage direction and camera orientation are flipped, which can be quite confusing when trying to figure out how the tiles are spatially related to each other. Do account for this and correctly register the data, we have to swap the xy axes. This can be accomplish by swap the last two axes of the tif,
 
@@ -107,4 +107,4 @@ Next, you'll run `04_pixeldecode_and_baysor.py` to first optimize the pixel-base
 
 ## Ensuring a sucessful run
 
-We have included the proper output of the `merfish3d-analysis` package for this dataset on the BIL servers in the `results` directory. You should be able to compare your obtained results to ours to ensure that there were no computation errors. One common issue we have run into is if the optical flow registration fails due to processor specific instructions (specifically older processors that lack AVX2). Please contact us if you have this issue, we have a workaround during installation that we can provide to you. 
+In progress.
